+++ unbabel/exceptions.py
@@ -1,0 +1,17 @@
+
+
+
+class UnauthorizedException(Exception):
+    def __init__(self, value):
+        self.value = value
+
+    def __str__(self):
+        return repr(self.value)
+
+
+class BadRequestException(Exception):
+    def __init__(self, value):
+        self.value = value
+
+    def __str__(self):
+        return repr(self.value)
